+++ cocoa-beans-commands/src/test/java/net/apartium/cocoabeans/commands/exception/SimpleArgumentMapperTest.java
@@ -7,8 +7,7 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 import java.util.Map;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.*;
 
 class SimpleArgumentMapperTest {
 
@@ -22,7 +21,16 @@ class SimpleArgumentMapperTest {
                 },
                 List.of(),
                 List.of(),
-                List.of(CommandException.class)
+                Map.of(CommandException.class, List.of(
+                        context -> context.parsedArgs().entrySet()
+                                .stream()
+                                .filter(entry -> CommandException.class.isAssignableFrom(entry.getKey()))
+                                .map(Map.Entry::getValue)
+                                .filter(list -> !list.isEmpty())
+                                .findFirst()
+                                .map(List::getFirst)
+                                .orElse(null)
+                ))
         );
 
         assertEquals(1, argumentIndices.size());
@@ -32,6 +40,49 @@ class SimpleArgumentMapperTest {
 
         assertEquals(commandException, argumentIndex.get(new ArgumentContext("test", new String[0], null, Map.of(CommandException.class, List.of(commandException)))));
         assertNull(argumentIndex.get(new ArgumentContext("test", new String[0], null, Map.of())));
+    }
+
+    @Test
+    void providingCustomIndex() {
+        ArgumentMapper mapper = new SimpleArgumentMapper();
+
+        Map<Class<?>, List<ArgumentIndex<?>>> providedArgumentIndexesByType = Map.of(String.class, List.of(context -> "this is cool", context -> "second value"));
+
+        List<ArgumentIndex<?>> argumentIndices = mapper.mapIndices(
+                new RegisteredVariant.Parameter[]{
+                        new RegisteredVariant.Parameter(String.class, String.class, new ArgumentRequirement[0], null)
+                },
+                List.of(),
+                List.of(),
+                providedArgumentIndexesByType
+        );
+
+        assertEquals(1, argumentIndices.size());
+        ArgumentIndex<?> argumentIndex = argumentIndices.get(0);
+
+        Object obj = argumentIndex.get(new ArgumentContext("test", new String[0], null, Map.of()));
+        assertNotNull(obj);
+
+        assertInstanceOf(String.class, obj);
+        assertEquals("this is cool", obj);
+
+        argumentIndices = mapper.mapIndices(
+                new RegisteredVariant.Parameter[]{
+                        new RegisteredVariant.Parameter(String.class, String.class, new ArgumentRequirement[0], null),
+                        new RegisteredVariant.Parameter(String.class, String.class, new ArgumentRequirement[0], null)
+                },
+                List.of(),
+                List.of(),
+                providedArgumentIndexesByType
+        );
+
+        assertEquals(2,  argumentIndices.size());
+
+        argumentIndex = argumentIndices.get(0);
+        assertEquals("this is cool", argumentIndex.get(new ArgumentContext("test", new String[0], null, Map.of())));
+
+        argumentIndex = argumentIndices.get(1);
+        assertEquals("second value",  argumentIndex.get(new ArgumentContext("test", new String[0], null, Map.of())));
     }
 
 }